buck2: don't access the `CommandBuilderContext` via `CommandLineBuilder`
Summary:
The CommandBuilder is what we accumulate in. The context allows us to produce
the paths we accumulate. There's no reason for the former to carry the latter,
which leads to an explosion in ways to do things (e.g. `ArgBuilder` is
notionally `CommandBuilder` but ... different).

Just pass those two distinct things as distinct things.

This boils down to making `add_to_command_line` take both the CLI to accumulate into, and the context.

Note: in a later diff, I'm going to rename `BaseCommandLineBuilder` to just be `DefaultCommandLineBuilderContext`, etc.

Reviewed By: ndmitchell

Differential Revision: D41478270

fbshipit-source-id: 89786d1eedfd606617e1e87fd5363a47a75097c9