Store only latest of each rdep
Summary:
we can avoid storing all versions of rdeps. Just store the lastest version of each.

This isn't very aggresive gc as it only clears upon the next computation.
We can technically GC on `commit` of version V, where V > current_version, but that's harder to do, so this is probably good enough for now.

Reviewed By: krallin

Differential Revision: D41292259

fbshipit-source-id: 264e85dd4380e863ab74174707fa414da02c18ed