Add link group link info for non-root nodes
Summary:
This updates `get_filtered_labels_to_links_map()` to add the corresponding
link group lib to the link line of a dependent, if the dependent references a
non-"root" node in that link group.  This unblocks auto-generating link group
libs, which don't correspond to a target in the linkable graph.

Reviewed By: christycylee

Differential Revision: D41589684

fbshipit-source-id: 83c3d8aaab2a2d8d363d39d46d99526282ce746c