Add a `sources` sub-target that only outputs the symlinked sources
Summary:
This changes adds a small `sources` sub-target to rust_library and rust_binary base rule to only exposes the directory containing the sources.
We're not guaranteed that the sources are being materialized when doing a regular buck building, which is fine. And we may also don't want to trigger the whole build just to inspect how the files are layed out.

Few use-cases:
- For rust-analyzer (see above diff in stack) (or I guess other IDEs), since it works often out of the box if it can read the source hierarchy.
- Anyone willing to check what's the output of a code-gen rule (like the crate generated by thrift libraries or bpf skeletons) without digging in the file system.
- Not willing to trigger full builds to check what sources would get used for a given build.

I may have a missed a hidden impact, but it seems like a light-enough proposal with no(?) downsides and a few QOL improvements

Reviewed By: jsgf, emersonford

Differential Revision: D41708907

fbshipit-source-id: a34d82607224085c69db55f4aebeeb482ba56aa1