cas_artifact for dirs
Summary:
This diff is a follow-up to D40947946 (https://github.com/facebookincubator/buck2/commit/dc5f41c20da5584fff3331b66b74f4eaae008ff9), which introduced `is_tree` to `cas_artifact`.

This diff adds in support for the *other* kind of CAS directory storage: `directory`s.

The difference between the two is that a `tree` is a self-contained collection of all of the file structure data, while a `directory` is just a root entry, with references to the other nodes requiring a CAS lookup. This means that accepting this as an artifact requires recursively downloading the metadata tree.

I implement that exact process in `re_directory_to_re_tree`, and the rest of the diff is just plumbing.

Reviewed By: krallin

Differential Revision: D41796325

fbshipit-source-id: 4b68550044c5a11a19ca77f581413502a9de548f