buck2: materializer: deferred: track queue size
Summary:
It's a useful thing to be able to observe, and I've been doing a lot of changes
here so I'd like to be able to see it when I benchmark.

Note that this also changes things for us to leak a few pointers for the
sender. In practice the materializer lives as long as the process so there is
no value in refcounting the senders & counters.

Reviewed By: ndmitchell

Differential Revision: D41801267

fbshipit-source-id: 0a55d85637229c90e2da33b91fcd5efb1c662b31