buck2: materializer: deferred: set `synchronous = OFF` for sqlite
Summary:
Setting synchronous to anything but OFF prevents data corruption in case of
power loss, but for the deferred materializer state, we are rather happy to run
the risk of data corruption (which we recover from by just dropping the state
and pretending we have none), rather than running a `fsync` at any point during
a build, which tends to be *very* slow, because if we do a `fsync` from the
deferred materializer, that will tend to occur after a lot of writes have been
done, which means a lot of data needs syncing.

Note that there isn't really a guarantee of ordering of things written across
different files anyway, so we always run some risk of having our state be
incorrect when power loss happens, unless we fsync after every single write,
but, huh, yeah, we aren't gonna be able to do that.

This problem manifests itself when routing writes through the deferred
materializer on a benchmark build. This causes the set of operations done by
the deferred materializer to exceed the WAL max size (which is 1000 pages that
are 4KB each, so about 4MB of data), which causes SQLite to write the WAL to
the database file, which is the only circumstance under which it'll do a
`fsync`, and we stall I/O.

Reviewed By: scottcao

Differential Revision: D41774541

fbshipit-source-id: 3fd9eec607d60cb9360b094b8e18724b33ba88c4