buck2: materializer: skip spawning tasks on clean when not needed
Summary:
Read here for why I'm doing this stack and for benchmarks: D41774542

 ---

Tokio task handles are actually non-trivially expensive when you have a lot of
them. In most cases, we don't need a Tokio task here (we only need one if we
have in-flight futures running in a path below the one being deleted, which is
fairly uncommon), so skip creating one.

Reviewed By: scottcao

Differential Revision: D41774548

fbshipit-source-id: c325ee5bbd4771b623d4ae4b3f7716825f2e58bc