add platform_deps_aliases for cxx_toolchains
Summary:
Infer is using `--target-platforms ovr_config//platform/linux:x86_64-platform010-infer` flag to select the infer toolchain.

This non-standard toolchain name makes some external_deps not matching (and thus not getting included) in the build properly.

For example; `third-party-buck/platform010/build/cuda`

```
7      exported_platform_deps = [
8          [
9              "^platform010-(?!clang-cuda)",
10             ["//third-party-buck/platform010/build/cuda/11.4.2:cuda"],
11         ],
12         [
```

With this change, a toolchain define a list of aliases that is used to resolve the platform_deps, and thus making it possible for infer to get all external deps as needed.

Reviewed By: krallin

Differential Revision: D41800486

fbshipit-source-id: 525a502be9a252d522dd3a1a1a2bdbf2691a415a