Distributed thinlto: make dist_lto_opt.py handle all flags and fix a missing flag issue for @fbcode/mode/opt-hhvm-lto
Summary:
## Background:
we found flag `"-hot-callsite-threshold=12000"` is missing from hhvm build. Because it's added through the `gen_modes.py`, and in such case it's not passed by `link_info`, we need to also take `cxx_link_cmd(ctx)` into account for opt phase.

To support such case, we need to move the current flag filter from `dist_lto.bzl` to `dist_lto_opt.py`, Which is the right thing to do and in our todo list, but now we need to prioritize it.

## Original summary from apolloww:
1. move filter and conversion logic to dist_lto_opt.py
2. dump pre/post-flags into an argsfile (*.opt.argsfile), so they are separated from those pre-set flags from `cxx_toolchain.cxx_compiler_info.compiler`.
3. extend dist_lto_opt.py argsparser to take input, index and split-dwarf values.
4. add unit test.

## V5:
1. Adding cmd args to the args file for opt phase `opt_args.add(cxx_link_cmd(ctx))`
2. buckify the unit tests, now it can be manually run with cmd `buck2 test cxx/dist_lto/tools:test_dist_lto_opt --target-platforms ovr_config//platform/linux:x86_64-fbcode-platform010-clang-nosan` under `fbcode/buck2/prelude/`
3. fixed an issue that default FDO profile gone missing during the the RE job.
4. add more unit tests

Reviewed By: apolloww

Differential Revision: D40961152

fbshipit-source-id: c30642bdadb08b5bf880d22c51d07666653a76a6