Daemon-side implementation of error categorization on LoadPackage step
Summary:
This Diff provides a structured erroring framework. The daemon constructs an error report that looks
```
message ErrorReport {
  optional ErrorCategory category = 1;         // { USER | INFRA }
  optional ErrorCause cause = 2;               // { INVALID_PACKAGE | ... }
  optional string error_message = 3;
}
```
and sends it to cli. Cli processes the data and then it can e.g. generate exit code and send a formatted log to Scuba, etc.

In code level, the error classification is done in-line by `anyhow::context` tagging, for example
```
        let root_entries = self
            .fs
            .read_dir(root.as_cell_path())
            .await
            .context(buck2_data::ErrorCategory::User)?;
```
When an undefined error occurs, the daemon will construct the error report with `ErrorCategory = INFRA` and `None` for the other fields (i.e. unknown) by default so that we can continue to iterate implementing more error categories. For example, if we see a high rate of `INFRA` errors at some step and with the same error message, we can attach a proper error tag to the line of code.

Only daemon-side implementation is done for this diff. Client-side one will follow in the next diff. Also, it only covers a few user errors such as "non-existing package dir speficied", "no permission to read the specified dir", "no build file exists", etc in only `LoadPackage` step for now.

----
# Why / Why Not
#### Why tagging?
Without tagging mechanism, we could instead implement a function like below to infer the error class.
```
/// Creates an `ErrorReport` based on `error`. The categorizing logic varies
/// according to which `step` the error occurred in.
pub fn categorize(
    error: &Option<anyhow::Error>,
    step: buck2_data::ErrorStep,
) -> Option<buck2_data::ErrorReport> {
...
```
Then categorization could possibly be done by regex-ing the string inside the `error`, but it's fragile since the error message is arbitral. It could also be done by trying to downcast `error` into several concrete types and infer the category, but in many cases the root cause contains only a very generic error (e.g. during LoadPackage step, `io::Error` occurs not only when the user specified a non-existing dir path (user error) but also when Eden is in trouble on the way (infra error,) so it would be difficult to tell the error category accurately. As such, the tagging is one of the good approaches. The tag itself gives the information of the cause and category in-place, so no independent categorizer function is needed.

#### Why anyhow::Context?
If we try to make our own error-propagating mechanism, it'll highly likely end up with implementing a very similar struct to anyhow::Context, so why not utilizing it? In addition, it has a good compatibility with error handling where we can write code cleanly, attaching a tag within one liner of early return of `?` operator, for example.

#### Why categorizing on daemon side?
Protobuf cannot convey error structures as they are, which are necessary to do classification work. (FYI currently, just a string (`"{:#}, err"`) is passed to the client when a daemon command fails.)

Reviewed By: krallin

Differential Revision: D41566342

fbshipit-source-id: 96eac469eea41ae9c20377c3bc0b6774846bb64e