buck2: SpanTracker: declare whether spans should be shown, not whether they are roots
Summary:
Currently, we declare which spans should be roots in the span tracker. This
is kind of annoying because we have to re-encode the hierarchy of which event
has which parent over in the SpanTracker.

This can result in mistakes, e.g. Buck2 currently shows 2 spans when you are
syncing DICE state + Waiting on Watchman, but in reality the latter in a child
of the former.

A better way to implement this is to indicate which spans should be rendered
to the user, then let the SpanTracker render them as roots if their parent
either does not exist or isn't rendered, or as children otherwise.

Note that this removes a few assertions in the SpanTracker (namely that parents
always start before their children and are removed after), but we've literally
never had a bug caused by violating those so they seem not all that valuable.

Reviewed By: ndmitchell

Differential Revision: D41994819

fbshipit-source-id: c62c66095e573ae6f5479003dbd403a8fb7df754