relocatable pars with remote modules
Summary:
This diff sketches out a new package style: `relocatable` packages. They are essentially `inplace` packages with an extra step at the end which uploads the entire "link-tree" to a remote store. The `par` entry point is changed to fetch the tree from this store before executing.

This means you can copy the `par` anywhere between machines and you'll be able to run it, but there's no cost of zipping and unzipping (instead we pay a download/upload cost).

NOTE: **DO NOT USE THIS IN PRODUCTION**; this implementation relies on `/usr/local/bin/frecli` for uploading and downloading from CAS so the output `par`s will only execute on machines with that binary.

Reviewed By: krallin

Differential Revision: D41654444

fbshipit-source-id: 0b680d09b3dbf8fc84ec1af641934b18d54833c9