buck2: simpleconsole: don't print all actions at default verbosity
Summary:
Like it says in the title. We'd received multiple reports that this is too much
output. This shifts action outputs to `-v`.

Note that we used to suggest using `-v` to get command lines, but in the
previous diffs I fixed this to suggest using `buck2 log what-failed` instead,
so this mostly "frees up" `-v` to be the actually "somewhat verbose" level,
instead of the "just give me the commands that failed" level.

Reviewed By: bobyangyf

Differential Revision: D42036220

fbshipit-source-id: dcd7d6f2c6d9e2d3fb1c09fc365ede07004d757c