Disallow '= None' on some attrs of buildscript_genrule_{args,srcs}
Summary: It doesn't make sense to me to allow passing in an explicit `cfgs = None` or `features = None` or `env = None` into `buildscript_genrule_args` or `buildscript_genrule_srcs`. Either do not pass it (D42178981) or pass an argument of the required type. We don't need `or []` / `or {}` throughout all the code downstream of the call.

Reviewed By: zertosh

Differential Revision: D42178982

fbshipit-source-id: 35105d9102cc460b61292749e04d2bc119a03215