Use actions_identifier instead of actions_prefix
Summary:
`actions.run` has an identifier which can be used when the same `category` is being used multiple times per target.

Right now for Java compilation, we are instead changing the `category`, which isn't really the best approach as we end up with a load of categories that only appear once and are only identifiable by the actions_prefix, e.g. see the bottom of this scuba table: https://fburl.com/scuba/buck2_actions/84q2txpo

This cleans it up to use `category` + `identifier`.

Reviewed By: navidqar

Differential Revision: D42139784

fbshipit-source-id: d4bfbfb7ef6ba628d166aa69b593a5203f725149