buck2: move getting build info to its own crate
Summary:
We don't use the regular build info to get the hg commit hash since that's
not working on macOS at the moment, so let's do that consistently everywhere
we use it. This means we'll get that info into our remote logging as well.

This also lets us deduplicate the logic for OSS-vs-not-OSS.

Reviewed By: ndmitchell

Differential Revision: D42306486

fbshipit-source-id: 880d49377a9be807dfac241fa3c305e6f56cbce1