buck2: fix chrome-trace
Summary:
This has been broken for a little while, because the Watchman span is a
child of the "graph sync" span, and that span isn't on the chrome trace.

So, rather than try to reproduce the hierarchy of spans here, just assume that
if a parent span is missing, that means we don't care to show it.

Reviewed By: IanChilds

Differential Revision: D42383247

fbshipit-source-id: 84d055e309e09ba7cb0dd330b50fb46e988600f8