Change the proto to include some changes and a reason
Summary: Before if there were 1000 file changes, we sent 0 of them, and just said there were too many. That isn't a great design, especially if you are talking to users - you probably want to know 10 that changed, plus a message that there were more. Change the proto to support that.

Reviewed By: bobyangyf

Differential Revision: D42142920

fbshipit-source-id: 6ad4bb71394ec11106a324263ea072301035e924