Selectify fbcode default link-style
Summary:
In v2, make the default link style func used in the macros a select,
and fixup handling of it throughout, to unblock making link style be
controlled by constraints on the target configuration (and unblock
making build mode be implemented via constraints).

This change should (mostly) be a noop.

Reviewed By: itamaro

Differential Revision: D42456152

fbshipit-source-id: bdb34bb7d3ceb40be5759b33e29bc219dcfb1407