Insert scribe write stats to snapshot events
Summary:
Building on top of the last diff, this diff reports scribe sink stats in buck2 snapshot events. Thankfully most of the plumbing exists between the `EventDispatch` struct containing a handle to scribe and the heartbeat guard that triggers snapshot event collection, so this is really more about defining what stats we want to capture.

Right now only scribe events are captured, but conceivably we could do the same for other sinks where it makes sense.

In situations where we have multiple sinks (e.g. with `TeeSink`) aggregation is done - right now just summing up counters. We could expand this in the future to splitting out metrics by sink if it matters, but we dont' have a compelling use case for other sinks so I'll defer this to later.

We emit the following metrics:
* Number of successful message sends
* Number of failures (that we can retry)
* Number of failures (that we won't retry)
* Number of attmepts we've tried to submit data
* Length of the internal buffer (how many messages we're storing to be processed)

We can expand these later on as needed.

Reviewed By: krallin

Differential Revision: D42443475

fbshipit-source-id: 7eaf90f8cef055b1a9a518796c43abc244c3f3a5