Actually flush scribe queue when requested
Summary:
As pointed on in D42421577 (https://github.com/facebookincubator/buck2/commit/1604bb9b6a14940426fd96393a21d7530d195bf7), with the changes in D42178265 we no longer actually flush the queue when requested.

This implements proper flushing of the queue. Note that this could still be a bit racy - nothing stops queue writers from adding more to the queue as `flush_blocking()` is running. Though races are generally bad, the amount of events we have will be bounded by the end of a buck2 invocation so we'll reach the end sometime (or timeout while trying to flush everything, or be ratelimited by scribe).

Reviewed By: krallin

Differential Revision: D42469080

fbshipit-source-id: a629926e7ae2082ab4cce8b5af10f3204de084a2